Change 'move workspace' feature to 'duplicate', delete the old duplicate workspace modal.
