+++ packages/insomnia/src/ui/components/dropdowns/workspace-card-dropdown.tsx
@@ -22,7 +22,6 @@ import { showError, showPrompt } from '../modals';
 import { ExportRequestsModal } from '../modals/export-requests-modal';
 import { ImportModal } from '../modals/import-modal';
 import { WorkspaceDuplicateModal } from '../modals/workspace-duplicate-modal';
-import { WorkspaceMoveModal } from '../modals/workspace-move-modal';
 import { WorkspaceSettingsModal } from '../modals/workspace-settings-modal';
 import { SvgIcon } from '../svg-icon';
 
@@ -83,10 +82,9 @@ const useDocumentActionPlugins = ({ workspace, apiSpec, project }: Props) => {
 };
 
 export const WorkspaceCardDropdown: FC<Props> = props => {
-  const { workspace, mockServer, project, projects } = props;
+  const { workspace, mockServer, project } = props;
   const fetcher = useFetcher();
   const [isDuplicateModalOpen, setIsDuplicateModalOpen] = useState(false);
-  const [isMoveModalOpen, setIsMoveModalOpen] = useState(false);
   const [isImportModalOpen, setIsImportModalOpen] = useState(false);
   const [isExportModalOpen, setIsExportModalOpen] = useState(false);
   const [isSettingsModalOpen, setIsSettingsModalOpen] = useState(false);
@@ -112,16 +110,6 @@ export const WorkspaceCardDropdown: FC<Props> = props => {
           </Button>
         }
       >
-        {/* Do not show move option for mock because mock url changes after move */}
-        {workspace.scope !== WorkspaceScopeKeys.mockServer && (
-          <DropdownItem aria-label='Move'>
-            <ItemContent
-              label="Move"
-              icon="arrow-right-arrow-left"
-              onClick={() => setIsMoveModalOpen(true)}
-            />
-          </DropdownItem>
-        )}
         <DropdownItem aria-label='Duplicate'>
           <ItemContent
             label="Duplicate"
@@ -199,19 +187,10 @@ export const WorkspaceCardDropdown: FC<Props> = props => {
           </DropdownItem>
         </DropdownSection>
       </Dropdown>
-      {isMoveModalOpen && (
-        <WorkspaceMoveModal
-          onHide={() => {
-            setIsMoveModalOpen(false);
-          }}
-          workspace={workspace}
-        />
-      )}
       {isDuplicateModalOpen && (
         <WorkspaceDuplicateModal
           onHide={() => setIsDuplicateModalOpen(false)}
           workspace={workspace}
-          projects={projects}
         />
       )}
       {isImportModalOpen && (
